+++ packages/code-studio/src/styleguide/DraggableLists.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/jsx-props-no-spreading */
 /* eslint no-console: "off" */
 import React, { Component } from 'react';
 import {
@@ -10,7 +9,7 @@ import {
 import memoize from 'memoizee';
 import { DragUtils, DraggableItemList, Range } from '@deephaven/components';
 import DraggableListInput from './DraggableListInput';
-import { sampleSectionIdAndClasses } from './utils';
+import SampleSection from './SampleSection';
 
 const DRAG_LIST_TITLES = ['Draggable Only', 'Drag and Drop', 'Droppable Only'];
 const DRAG_LIST_PROPS = [
@@ -190,11 +189,7 @@ class DraggableLists extends Component<
   render(): React.ReactElement {
     const { items, lists, selectedRanges } = this.state;
     return (
-      <div
-        {...sampleSectionIdAndClasses('draggable-lists', [
-          'style-guide-inputs',
-        ])}
-      >
+      <SampleSection name="draggable-lists" className="style-guide-inputs">
         <h2 className="ui-title">Drag and Drop Lists</h2>
         <div className="row">
           <DragDropContext
@@ -219,7 +214,7 @@ class DraggableLists extends Component<
             ))}
           </DragDropContext>
         </div>
-      </div>
+      </SampleSection>
     );
   }
 }