+++ packages/code-studio/src/styleguide/SpectrumComparison.tsx
@@ -1,6 +1,5 @@
-/* eslint-disable react/jsx-props-no-spreading */
 /* eslint-disable react/style-prop-object */
-import React, { Fragment, useCallback, useState } from 'react';
+import React, { Fragment, useState } from 'react';
 import {
   ActionButton,
   Button,
@@ -11,8 +10,6 @@ import {
   Icon,
   Item,
   Picker,
-  Radio,
-  RadioGroup,
   SpectrumButtonProps,
   TextField,
 } from '@adobe/react-spectrum';
@@ -20,8 +17,6 @@ import {
   Button as BootstrapButtonOld,
   Checkbox as CheckboxOld,
   ComboBox as ComboBoxOld,
-  RadioGroup as RadioGroupOld,
-  RadioItem,
   Select,
   View,
   Text,
@@ -33,7 +28,7 @@ import {
   SAMPLE_SECTION_E2E_IGNORE,
   SPECTRUM_COMPARISON_SAMPLES_ID,
 } from './constants';
-import { sampleSectionIdAndClasses } from './utils';
+import SampleSection from './SampleSection';
 
 type BootstrapLevel = 'primary' | 'secondary' | 'danger';
 
@@ -51,20 +46,11 @@ const options = [
 
 export function SpectrumComparison(): JSX.Element {
   const [isChecked, setIsChecked] = useState(false);
-  const [radioValue, setRadioValue] = useState('1');
-
-  const handleRadioChange = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => {
-      setRadioValue(event.target.value);
-    },
-    []
-  );
 
   return (
-    <div
-      {...sampleSectionIdAndClasses(SPECTRUM_COMPARISON_SAMPLES_ID, [
-        SAMPLE_SECTION_E2E_IGNORE,
-      ])}
+    <SampleSection
+      name={SPECTRUM_COMPARISON_SAMPLES_ID}
+      className={SAMPLE_SECTION_E2E_IGNORE}
     >
       <h2 className="ui-title" data-no-menu>
         Bootstrap / Spectrum Comparison
@@ -285,31 +271,10 @@ export function SpectrumComparison(): JSX.Element {
               Disabled
             </CheckboxOld>
             <Checkbox isDisabled>Disabled</Checkbox>
-
-            <Flex direction="column">
-              <label>
-                Radio Group
-                <RadioGroupOld onChange={handleRadioChange} value={radioValue}>
-                  <RadioItem value="1">One</RadioItem>
-                  <RadioItem value="2">Two</RadioItem>
-                  <RadioItem value="3">Three</RadioItem>
-                </RadioGroupOld>
-              </label>
-            </Flex>
-
-            <RadioGroup
-              label="Radio Group"
-              value={radioValue}
-              onChange={setRadioValue}
-            >
-              <Radio value="1">One</Radio>
-              <Radio value="2">Two</Radio>
-              <Radio value="3">Three</Radio>
-            </RadioGroup>
           </Grid>
         </View>
       </Flex>
-    </div>
+    </SampleSection>
   );
 }