+++ packages/components/src/spectrum/forms.ts
@@ -1,13 +1,14 @@
 export {
-  // Checkbox - we want to use our own `Checkbox` component instead of Spectrum's
+  Checkbox as SpectrumCheckbox,
+  type SpectrumCheckboxProps,
   CheckboxGroup,
   type SpectrumCheckboxGroupProps as CheckboxGroupProps,
   Form,
   type SpectrumFormProps as FormProps,
   NumberField,
   type SpectrumNumberFieldProps as NumberFieldProps,
-  // Radio - re-export once we re-export RadioGroup
-  // RadioGroup - we need to replace references to our DH `RadioGroup` before we can re-export
+  Radio,
+  RadioGroup,
   RangeSlider,
   type SpectrumRangeSliderProps as RangeSliderProps,
   Slider,