+++ packages/dashboard/src/PanelManager.ts
@@ -362,9 +362,11 @@ class PanelManager {
         config.component,
         config
       );
-      (dehydratedConfig as ClosedPanel).parentStackId =
-        LayoutUtils.getStackForConfig(root, config)?.config.id;
-      this.closed.push(dehydratedConfig);
+      if (dehydratedConfig != null) {
+        (dehydratedConfig as ClosedPanel).parentStackId =
+          LayoutUtils.getStackForConfig(root, config)?.config.id;
+        this.closed.push(dehydratedConfig);
+      }
     }
   }