+++ tests/utils.ts
@@ -7,7 +7,6 @@ import {
   webkit,
 } from '@playwright/test';
 import os from 'node:os';
-import shortid from 'shortid';
 
 export const HIDE_FROM_E2E_TESTS_CLASS = 'hide-from-e2e-tests';
 
@@ -126,6 +125,19 @@ export async function openPlot(
 }
 
 /**
+ * Generate a unique Id
+ * @param length Length to give id
+ * @returns A unique valid id
+ */
+export function generateId(length = 21): string {
+  let id = '';
+  for (let i = 0; i < length; i += 1) {
+    id += Math.random().toString(36).substr(2, 1);
+  }
+  return id;
+}
+
+/**
  * Generate a unique python variable name
  * @param prefix Prefix to give the variable name
  * @returns A unique string that is a valid python variable name
@@ -134,7 +146,7 @@ export function generateVarName(prefix = 'v'): string {
   // Don't allow a `-` in variable names
   let id: string;
   do {
-    id = shortid();
+    id = generateId();
   } while (id.includes('-'));
   return `${prefix}_${id}`;
 }